+++ NEWS.md
@@ -3,15 +3,14 @@
 
 ## v0.3.0
 
-- Package extensions for Dates.jl, LaTeXStrings.jl, and Markdown.jl
-
 ### Strings
 
 - New default context `backticks = 3`
 - The default value for the `indent` context is now `typst"  "`, to correspond with Typst's default indentation
 - Typst parameters are now printed on their own lines using the `indent` and `depth` settings
 - `show_typst`
-    - An `AbstractString` is no longer quoted and escaped in `code` and `math` mode
+    - `Signed` in `markup` mode is now converted to a `code` mode integer
+    - An `AbstractChar` and `AbstractString` now correspond to a Typst string
     - Implemented for `Dates.Date`, `Dates.DateTime`, `Dates.Day`, `Dates.Hour`, `Dates.Minute`, `Dates.Second`, `Dates.Time`, `Dates.Week`, `Docs.HTML`, `LaTeXStrings.LaTeXString`, `Markdown.MD`, and `VersionNumber`
 - The `show` methods for `TypstString` with a `application/pdf`, `image/png`, and `image/svg+xml` MIME types now support a custom `preamble`